requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-jsonfield-backport==1.0.5
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
